Rename PhoneField components
